[type:feat] modify sdk starter and add ShenyuInstanceRegisterRepository bean .
